requirements.txt
ckanapi==4.9
simplejson==3.20.2
importlib_metadata; python_version < "3.8"
